Add button for move between login and register
